Update Camel to 4.4.1
- based on camel-kamelets 4.4.1
- Update camel-k-crds to 2.3.0
