+++ library/camel-kamelets-catalog/src/test/java/org/apache/camel/kamelets/catalog/KameletsCatalogTest.java
@@ -146,17 +146,17 @@ public class KameletsCatalogTest {
         verifyHeaders("azure-storage-queue-source", 6);
         verifyHeaders("cassandra-sink", 1);
         verifyHeaders("cassandra-source", 1);
-        verifyHeaders("elasticsearch-index-sink", 9);
+        verifyHeaders("elasticsearch-index-sink", 10);
         verifyHeaders("ftp-source", 10);
         verifyHeaders("ftp-sink", 8);
         verifyHeaders("http-sink", 14);
         verifyHeaders("jira-add-comment-sink", 17);
         verifyHeaders("jira-add-issue-sink", 17);
         verifyHeaders("jira-source", 3);
-        verifyHeaders("jms-amqp-10-source", 14);
-        verifyHeaders("jms-amqp-10-sink", 17);
-        verifyHeaders("jms-ibm-mq-source", 14);
-        verifyHeaders("jms-ibm-mq-sink", 17);
+        verifyHeaders("jms-amqp-10-source", 15);
+        verifyHeaders("jms-amqp-10-sink", 18);
+        verifyHeaders("jms-ibm-mq-source", 15);
+        verifyHeaders("jms-ibm-mq-sink", 18);
         verifyHeaders("kafka-source", 9);
         verifyHeaders("kafka-sink", 5);
         verifyHeaders("log-sink", 0);
